requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.7
Pillow==5.3.0
python-resize-image==1.1.18
requests==2.20.1
urllib3==1.24.1
